refactor: override deprecated DirWatcher with DirWatcher2
